patterned the UI for the services and their subscreens
