+++ lib/screens/Services.dart
@@ -4,6 +4,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:navbar_router/navbar_router.dart';
+import 'package:sport_test/subscreens/climb.dart';
+import 'package:sport_test/subscreens/gymclass.dart';
+import 'package:sport_test/subscreens/gymsesh.dart';
+import 'package:sport_test/subscreens/swim.dart';
 
 class ServiceNavigatorRoutes {
   static const String root = '/';
@@ -97,189 +101,147 @@ class ServiceState extends State<Service> {
         ),
         body: Center(
             child: GridView(
-                padding: const EdgeInsets.all(10),
+                padding: const EdgeInsets.all(20),
                 gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+                  mainAxisSpacing: 15,
                   crossAxisCount: 2,
                   crossAxisSpacing: 15.0,
                 ),
                 children: [
+              // GYM CARD
               InkWell(
                 onTap: () {
                   Navigator.of(context).push(
                       MaterialPageRoute(builder: (context) => const Gymsesh()));
                 },
                 child: Card(
-                  shape: const RoundedRectangleBorder(
-                    side: BorderSide(
-                      color: Colors.white,
-                    ),
-                    borderRadius: BorderRadius.all(Radius.circular(5)),
-                  ),
                   color: Colors.purple,
-                  elevation: 20,
-                  borderOnForeground: true,
-                  // surfaceTintColor: Colors.purpleAccent,
-                  shadowColor: Colors.white,
+                  shape: const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(15)),
+                  ),
                   child: Column(
-                    mainAxisAlignment: MainAxisAlignment.start,
                     children: [
-                      Image.asset(
-                        'lib/assets/images/ravelin_gym2.png',
-                        fit: BoxFit.cover,
-                        alignment: Alignment.center,
-                        // height: 50,
-                        // width: 50,
-                        // color: Colors.white,
+                      ClipRRect(
+                        borderRadius: const BorderRadius.only(
+                            topLeft: Radius.circular(15),
+                            topRight: Radius.circular(15)),
+                        child: Image.asset('lib/assets/images/ravelin_gym.png',
+                            height: 110,
+                            gaplessPlayback: true,
+                            fit: BoxFit.fill),
                       ),
                       const SizedBox(
                         height: 10,
                       ),
                       const Text(
                         'Gym',
-                        style: TextStyle(
-                            fontSize: 25,
-                            // decoration: TextDecoration.overline,
-                            // fontWeight: FontWeight.bold,
-                            color: Colors.white),
-                      ),
+                        style: TextStyle(fontSize: 20, color: Colors.white),
+                      )
                     ],
                   ),
                 ),
               ),
+
+              //CLASSES CARD
               InkWell(
                 onTap: () {
-                  navigate(
-                    context,
-                    Gymclass.route,
-                    isRootNavigator: false,
-                  );
+                  Navigator.of(context).push(MaterialPageRoute(
+                      builder: (context) => const Gymclass()));
                 },
                 child: Card(
-                  shape: const RoundedRectangleBorder(
-                    side: BorderSide(
-                      color: Colors.white,
-                    ),
-                    borderRadius: BorderRadius.all(Radius.circular(5)),
-                  ),
                   color: Colors.purple,
-                  elevation: 20,
-                  borderOnForeground: true,
-                  // surfaceTintColor: Colors.purpleAccent,
-                  shadowColor: Colors.white,
+                  shape: const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(15)),
+                  ),
                   child: Column(
-                    mainAxisAlignment: MainAxisAlignment.start,
                     children: [
-                      Image.asset(
-                        'lib/assets/images/ravelin_class.png',
-                        fit: BoxFit.cover,
-                        alignment: Alignment.center,
-                        // height: 50,
-                        // width: 50,
-                        // color: Colors.white,
+                      ClipRRect(
+                        borderRadius: const BorderRadius.only(
+                            topLeft: Radius.circular(15),
+                            topRight: Radius.circular(15)),
+                        child: Image.asset(
+                            'lib/assets/images/ravelin_class.png',
+                            height: 110,
+                            gaplessPlayback: true,
+                            fit: BoxFit.fill),
                       ),
                       const SizedBox(
                         height: 10,
                       ),
                       const Text(
                         'Classes',
-                        style: TextStyle(
-                            fontSize: 25,
-                            // decoration: TextDecoration.overline,
-                            // fontWeight: FontWeight.bold,
-                            color: Colors.white),
-                      ),
+                        style: TextStyle(fontSize: 20, color: Colors.white),
+                      )
                     ],
                   ),
                 ),
               ),
+
+              //SWIM CARD
               InkWell(
                 onTap: () {
-                  navigate(
-                    context,
-                    Swim.route,
-                    isRootNavigator: false,
-                  );
+                  Navigator.of(context).push(
+                      MaterialPageRoute(builder: (context) => const Swim()));
                 },
                 child: Card(
-                  shape: const RoundedRectangleBorder(
-                    side: BorderSide(
-                      color: Colors.white,
-                    ),
-                    borderRadius: BorderRadius.all(Radius.circular(5)),
-                  ),
                   color: Colors.purple,
-                  elevation: 20,
-                  borderOnForeground: true,
-                  // surfaceTintColor: Colors.purpleAccent,
-                  shadowColor: Colors.white,
+                  shape: const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(15)),
+                  ),
                   child: Column(
-                    mainAxisAlignment: MainAxisAlignment.start,
                     children: [
-                      Image.asset(
-                        'lib/assets/images/ravelin_swim.png',
-                        fit: BoxFit.cover,
-                        alignment: Alignment.center,
-                        // height: 50,
-                        // width: 50,
-                        // color: Colors.white,
+                      ClipRRect(
+                        borderRadius: const BorderRadius.only(
+                            topLeft: Radius.circular(15),
+                            topRight: Radius.circular(15)),
+                        child: Image.asset('lib/assets/images/ravelin_swim.png',
+                            height: 110, width: 180, fit: BoxFit.fill),
                       ),
                       const SizedBox(
                         height: 10,
                       ),
                       const Text(
                         'Swim',
-                        style: TextStyle(
-                            fontSize: 25,
-                            // decoration: TextDecoration.overline,
-                            // fontWeight: FontWeight.bold,
-                            color: Colors.white),
-                      ),
+                        style: TextStyle(fontSize: 20, color: Colors.white),
+                      )
                     ],
                   ),
                 ),
               ),
+
+              //CLIMB CARD
               InkWell(
                 onTap: () {
-                  navigate(
-                    context,
-                    Climb.route,
-                    isRootNavigator: false,
-                  );
+                  Navigator.of(context).push(
+                      MaterialPageRoute(builder: (context) => const Climb()));
                 },
                 child: Card(
                   color: Colors.purple,
-                  elevation: 20,
-                  borderOnForeground: true,
                   shape: const RoundedRectangleBorder(
-                    side: BorderSide(
-                      color: Colors.white,
-                    ),
-                    borderRadius: BorderRadius.all(Radius.circular(5)),
+                    borderRadius: BorderRadius.all(Radius.circular(15)),
                   ),
-                  surfaceTintColor: Colors.purpleAccent,
-                  shadowColor: Colors.white,
                   child: Column(
-                    mainAxisAlignment: MainAxisAlignment.start,
                     children: [
-                      Image.asset(
-                        'lib/assets/images/ravelin_climb.png',
-                        fit: BoxFit.cover,
-                        alignment: Alignment.center,
-                        height: 125,
-                        width: 250,
-                        // color: Colors.white,
+                      ClipRRect(
+                        borderRadius: const BorderRadius.only(
+                            topLeft: Radius.circular(15),
+                            topRight: Radius.circular(15)),
+                        child: Image.asset(
+                            'lib/assets/images/ravelin_climb.png',
+                            height: 110,
+                            gaplessPlayback: true,
+                            fit: BoxFit.fill),
                       ),
                       const SizedBox(
                         height: 10,
                       ),
                       const Text(
-                        'Wall Climb',
+                        'Climb',
                         style: TextStyle(
-                            fontSize: 25,
-                            // decoration: TextDecoration.overline,
-                            // fontWeight: FontWeight.bold,
-                            color: Colors.white),
-                      ),
+                            fontSize: 20,
+                            color: Colors.white,
+                            fontWeight: FontWeight.normal),
+                      )
                     ],
                   ),
                 ),
@@ -313,945 +275,5 @@ class ActionButton extends StatelessWidget {
         color: Colors.white,
       ),
     );
-  }
-}
-
-//Gym session page
-
-class Gymsesh extends StatefulWidget {
-  static const String route = '/Bookings/Gymsesh';
-  const Gymsesh({Key? key}) : super(key: key);
-
-  @override
-  State<Gymsesh> createState() => GymseshState();
-}
-
-class GymseshState extends State<Gymsesh> {
-  @override
-  Widget build(BuildContext context) {
-    return WillPopScope(
-      onWillPop: () async {
-        Navigator.of(context)
-            .pop();
-        return true;
-      },
-      child: Scaffold(
-        appBar: AppBar(
-          title: const Text('Gym Session'),
-          centerTitle: true,
-        ),
-        body: GridView(
-          padding: const EdgeInsets.all(20),
-          // controller: swimscrollController,
-          gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-              crossAxisCount: 2),
-          children: [
-            InkWell(
-              onTap: () {
-                NavbarNotifier.hideBottomNavBar = false;
-                navigate(context, BookingCalendarView.route,
-                    isRootNavigator: false);
-              },
-              child: Card(
-                shape: const RoundedRectangleBorder(
-                  side: BorderSide(
-                    color: Colors.white,
-                  ),
-                  // borderRadius: BorderRadius.all(Radius.circular(2)),
-                ),
-                elevation: 20,
-                color: Colors.purple,
-                borderOnForeground: true,
-                surfaceTintColor: Colors.purple,
-                shadowColor: Colors.black,
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  children: [
-                    Image.asset(
-                      'lib/assets/images/gym_lifting.png',
-                      fit: BoxFit.fill,
-                      alignment: Alignment.center,
-                      height: 115,
-                      width: 250,
-                    ),
-                    const SizedBox(
-                      height: 5,
-                    ),
-                    const Text(
-                      'Gym',
-                      style: TextStyle(
-                          fontSize: 20,
-                          // fontWeight: FontWeight.bold,
-                          color: Colors.white),
-                    ),
-                    const Text(
-                      'Bookings  >',
-                      style: TextStyle(
-                          fontSize: 20,
-                          // fontWeight: FontWeight.bold,
-                          color: Colors.white),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            InkWell(
-              onTap: () {},
-              child: Card(
-                shape: const RoundedRectangleBorder(
-                  side: BorderSide(
-                    color: Colors.white,
-                  ),
-                  // borderRadius: BorderRadius.all(Radius.circular(12)),
-                ),
-                elevation: 20,
-                color: Colors.purple,
-                borderOnForeground: true,
-                surfaceTintColor: Colors.purple,
-                shadowColor: Colors.white,
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  children: [
-                    Image.asset(
-                      'lib/assets/images/Gym_booking.png',
-                      fit: BoxFit.fill,
-                      alignment: Alignment.center,
-                    ),
-                    const SizedBox(
-                      height: 8,
-                    ),
-                    const Text(
-                      'Gym ',
-                      style: TextStyle(
-                          fontSize: 20,
-                          // fontWeight: FontWeight.bold,
-                          color: Colors.white),
-                    ),
-                    const Text(
-                      'Services  >',
-                      style: TextStyle(
-                          fontSize: 20,
-                          // fontWeight: FontWeight.bold,
-                          color: Colors.white),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            InkWell(
-              onTap: () {},
-              child: Card(
-                shape: const RoundedRectangleBorder(
-                  side: BorderSide(
-                    color: Colors.white,
-                  ),
-                  // borderRadius: BorderRadius.all(Radius.circular(12)),
-                ),
-                elevation: 20,
-                color: Colors.purple,
-                borderOnForeground: true,
-                surfaceTintColor: Colors.purple,
-                shadowColor: Colors.white,
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  children: [
-                    Image.asset(
-                      'lib/assets/images/gym_services.png',
-                      fit: BoxFit.fill,
-                      alignment: Alignment.center,
-                    ),
-                    const SizedBox(
-                      height: 7,
-                    ),
-                    const Text(
-                      'Gym',
-                      style: TextStyle(
-                          fontSize: 20,
-                          // fontWeight: FontWeight.bold,
-                          color: Colors.white),
-                    ),
-                    const Text(
-                      '   rules >',
-                      style: TextStyle(
-                          fontSize: 20,
-                          // fontWeight: FontWeight.bold,
-                          color: Colors.white),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-//Gym Class Page
-class Gymclass extends StatefulWidget {
-  static const String route = '/Bookings/Gymclass';
-  const Gymclass({Key? key}) : super(key: key);
-
-  @override
-  State<Gymclass> createState() => GymclassState();
-}
-
-class GymclassState extends State<Gymclass> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Classes'),
-        centerTitle: true,
-      ),
-      body: GridView(
-        padding: const EdgeInsets.all(20),
-        // controller: swimscrollController,
-        gridDelegate:
-            const SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: 2),
-        children: [
-          InkWell(
-            onTap: () {},
-            child: Card(
-              shape: const RoundedRectangleBorder(
-                side: BorderSide(
-                  color: Colors.white,
-                ),
-                // borderRadius: BorderRadius.all(Radius.circular(2)),
-              ),
-              elevation: 20,
-              color: Colors.purple,
-              borderOnForeground: true,
-              surfaceTintColor: Colors.purple,
-              shadowColor: Colors.black,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Image.asset(
-                    'lib/assets/images/classes_3.png',
-                    fit: BoxFit.fill,
-                    alignment: Alignment.center,
-                    height: 115,
-                    width: 250,
-                  ),
-                  const SizedBox(
-                    height: 5,
-                  ),
-                  const Text(
-                    'Class',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                  const Text(
-                    'Bookings  >',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          InkWell(
-            onTap: () {},
-            child: Card(
-              shape: const RoundedRectangleBorder(
-                side: BorderSide(
-                  color: Colors.white,
-                ),
-                // borderRadius: BorderRadius.all(Radius.circular(12)),
-              ),
-              elevation: 20,
-              color: Colors.purple,
-              borderOnForeground: true,
-              surfaceTintColor: Colors.purple,
-              shadowColor: Colors.white,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Image.asset(
-                    'lib/assets/images/aquaclass_2.png',
-                    fit: BoxFit.fill,
-                    alignment: Alignment.center,
-                    height: 110,
-                    width: 168,
-                  ),
-                  const SizedBox(
-                    height: 8,
-                  ),
-                  const Text(
-                    'Aqua',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                  const Text(
-                    '   Classes  >',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          InkWell(
-            onTap: () {},
-            child: Card(
-              shape: const RoundedRectangleBorder(
-                side: BorderSide(
-                  color: Colors.white,
-                ),
-                // borderRadius: BorderRadius.all(Radius.circular(12)),
-              ),
-              elevation: 20,
-              color: Colors.purple,
-              borderOnForeground: true,
-              surfaceTintColor: Colors.purple,
-              shadowColor: Colors.white,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Image.asset(
-                    'lib/assets/images/classes_4.png',
-                    fit: BoxFit.fill,
-                    alignment: Alignment.center,
-                    height: 115,
-                    width: 250,
-                  ),
-                  const SizedBox(
-                    height: 7,
-                  ),
-                  const Text(
-                    'Class',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                  const Text(
-                    '  Rules >',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          InkWell(
-            onTap: () {},
-            child: Card(
-              shape: const RoundedRectangleBorder(
-                side: BorderSide(
-                  color: Colors.white,
-                ),
-                // borderRadius: BorderRadius.all(Radius.circular(12)),
-              ),
-              elevation: 20,
-              color: Colors.purple,
-              borderOnForeground: true,
-              surfaceTintColor: Colors.purple,
-              shadowColor: Colors.white,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Image.asset(
-                    'lib/assets/images/session_help.png',
-                    fit: BoxFit.fill,
-                    alignment: Alignment.center,
-                  ),
-                  const SizedBox(
-                    height: 8,
-                  ),
-                  const Text(
-                    'UOP',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                  const Text(
-                    'Fitness  >',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-//Climbing Page
-class Climb extends StatefulWidget {
-  static const String route = '/Bookings/Climb';
-  const Climb({Key? key}) : super(key: key);
-
-  @override
-  State<Climb> createState() => ClimbState();
-}
-
-class ClimbState extends State<Climb> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Climbing'),
-        centerTitle: true,
-      ),
-      body: GridView(
-        padding: const EdgeInsets.all(20),
-        // controller: swimscrollController,
-        gridDelegate:
-            const SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: 2),
-        children: [
-          InkWell(
-            onTap: () {},
-            child: Card(
-              shape: const RoundedRectangleBorder(
-                side: BorderSide(
-                  color: Colors.white,
-                ),
-                borderRadius: BorderRadius.all(Radius.circular(5)),
-              ),
-              elevation: 20,
-              color: Colors.purple,
-              borderOnForeground: true,
-              surfaceTintColor: Colors.purple,
-              shadowColor: Colors.black,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Image.asset(
-                    'lib/assets/images/climb_girl.png',
-                    fit: BoxFit.fill,
-                    alignment: Alignment.center,
-                    height: 115,
-                    width: 170,
-                  ),
-                  const SizedBox(
-                    height: 5,
-                  ),
-                  const Text(
-                    'Climb',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                  const Text(
-                    '  Bookings  >',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          InkWell(
-            onTap: () {},
-            child: Card(
-              shape: const RoundedRectangleBorder(
-                side: BorderSide(
-                  color: Colors.white,
-                ),
-                borderRadius: BorderRadius.all(Radius.circular(5)),
-              ),
-              elevation: 20,
-              color: Colors.purple,
-              borderOnForeground: true,
-              surfaceTintColor: Colors.purple,
-              shadowColor: Colors.white,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Image.asset(
-                    'lib/assets/images/climb_3.png',
-                    fit: BoxFit.fill,
-                    alignment: Alignment.center,
-                  ),
-                  const SizedBox(
-                    height: 8,
-                  ),
-                  const Text(
-                    'Session',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                  const Text(
-                    'Help  >',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          InkWell(
-            onTap: () {},
-            child: Card(
-              shape: const RoundedRectangleBorder(
-                side: BorderSide(
-                  color: Colors.white,
-                ),
-                borderRadius: BorderRadius.all(Radius.circular(5)),
-              ),
-              elevation: 20,
-              color: Colors.purple,
-              borderOnForeground: true,
-              surfaceTintColor: Colors.purple,
-              shadowColor: Colors.white,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Image.asset(
-                    'lib/assets/images/climb_1.png',
-                    fit: BoxFit.fill,
-                    alignment: Alignment.center,
-                    height: 115,
-                    width: 170,
-                  ),
-                  const SizedBox(
-                    height: 7,
-                  ),
-                  const Text(
-                    'Climbing',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                  const Text(
-                    ' Equipments >',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          InkWell(
-            onTap: () {},
-            child: Card(
-              shape: const RoundedRectangleBorder(
-                side: BorderSide(
-                  color: Colors.white,
-                ),
-                borderRadius: BorderRadius.all(Radius.circular(5)),
-              ),
-              elevation: 20,
-              color: Colors.purple,
-              borderOnForeground: true,
-              surfaceTintColor: Colors.purple,
-              shadowColor: Colors.white,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Image.asset(
-                    'lib/assets/images/climb_2.png',
-                    fit: BoxFit.fill,
-                    alignment: Alignment.center,
-                  ),
-                  const SizedBox(
-                    height: 8,
-                  ),
-                  const Text(
-                    'Under 16',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                  const Text(
-                    'Bookings  >',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-//Swim Page
-class Swim extends StatefulWidget {
-  static const String route = '/Bookings/Swim';
-  const Swim({Key? key}) : super(key: key);
-
-  @override
-  State<Swim> createState() => SwimState();
-}
-
-class SwimState extends State<Swim> {
-  final swimscrollController = ScrollController();
-
-  @override
-  void didChangeDependencies() {
-    super.didChangeDependencies();
-    size = MediaQuery.of(context).size;
-    if (size.width < 50) {
-      _addScrollListener();
-    }
-  }
-
-  void handleScroll() {
-    if (size.width > 50) return;
-    if (swimscrollController.position.userScrollDirection ==
-        ScrollDirection.forward) {
-      if (NavbarNotifier.isNavbarHidden) {
-        NavbarNotifier.hideBottomNavBar = false;
-      }
-    } else {
-      if (!NavbarNotifier.isNavbarHidden) {
-        NavbarNotifier.hideBottomNavBar = true;
-      }
-    }
-  }
-
-  void _addScrollListener() {
-    swimscrollController.addListener(handleScroll);
-  }
-
-  Size size = Size.zero;
-  @override
-  void dispose() {
-    swimscrollController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Swimming, steam and Sauna'),
-        centerTitle: true,
-      ),
-      body: GridView(
-        padding: const EdgeInsets.all(20),
-        controller: swimscrollController,
-        gridDelegate:
-            const SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: 2),
-        children: [
-          InkWell(
-            onTap: () {},
-            child: Card(
-              shape: const RoundedRectangleBorder(
-                side: BorderSide(
-                  color: Colors.white,
-                ),
-                // borderRadius: BorderRadius.all(Radius.circular(2)),
-              ),
-              elevation: 20,
-              color: Colors.purple,
-              borderOnForeground: true,
-              surfaceTintColor: Colors.purple,
-              shadowColor: Colors.black,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Image.asset(
-                    'lib/assets/images/man_swim.png',
-                    fit: BoxFit.fill,
-                    alignment: Alignment.center,
-                  ),
-                  const SizedBox(
-                    height: 5,
-                  ),
-                  const Text(
-                    'Swim',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                  const Text(
-                    'Bookings  >',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          InkWell(
-            onTap: () {},
-            child: Card(
-              shape: const RoundedRectangleBorder(
-                side: BorderSide(
-                  color: Colors.white,
-                ),
-                // borderRadius: BorderRadius.all(Radius.circular(12)),
-              ),
-              elevation: 20,
-              color: Colors.purple,
-              borderOnForeground: true,
-              surfaceTintColor: Colors.purple,
-              shadowColor: Colors.white,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Image.asset(
-                    'lib/assets/images/sauna.png',
-                    fit: BoxFit.fill,
-                    alignment: Alignment.center,
-                  ),
-                  const SizedBox(
-                    height: 8,
-                  ),
-                  const Text(
-                    'Steam &',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                  const Text(
-                    'Sauna  >',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          InkWell(
-            onTap: () {},
-            child: Card(
-              shape: const RoundedRectangleBorder(
-                side: BorderSide(
-                  color: Colors.white,
-                ),
-                // borderRadius: BorderRadius.all(Radius.circular(12)),
-              ),
-              elevation: 20,
-              color: Colors.purple,
-              borderOnForeground: true,
-              surfaceTintColor: Colors.purple,
-              shadowColor: Colors.white,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Image.asset(
-                    'lib/assets/images/under_16.png',
-                    fit: BoxFit.fill,
-                    alignment: Alignment.center,
-                  ),
-                  const SizedBox(
-                    height: 7,
-                  ),
-                  const Text(
-                    'Under 16',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                  const Text(
-                    ' Booking >',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-          InkWell(
-            onTap: () {},
-            child: Card(
-              shape: const RoundedRectangleBorder(
-                side: BorderSide(
-                  color: Colors.white,
-                ),
-                // borderRadius: BorderRadius.all(Radius.circular(12)),
-              ),
-              elevation: 20,
-              color: Colors.purple,
-              borderOnForeground: true,
-              surfaceTintColor: Colors.purple,
-              shadowColor: Colors.white,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.start,
-                children: [
-                  Image.asset(
-                    'lib/assets/images/session_help.png',
-                    fit: BoxFit.fill,
-                    alignment: Alignment.center,
-                  ),
-                  const SizedBox(
-                    height: 8,
-                  ),
-                  const Text(
-                    'Session',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                  const Text(
-                    'Help  >',
-                    style: TextStyle(
-                        fontSize: 20,
-                        // fontWeight: FontWeight.bold,
-                        color: Colors.white),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-//The booking view.
-class BookingCalendarView extends StatefulWidget {
-  static const String route = '/Bookings/Gymsesh/calendarView';
-  const BookingCalendarView({Key? key}) : super(key: key);
-
-  @override
-  State<BookingCalendarView> createState() => _BookingCalendarViewState();
-}
-
-class _BookingCalendarViewState extends State<BookingCalendarView> {
-  final now = DateTime.now();
-  late BookingService sportBookingService;
-
-  @override
-  void initState() {
-    super.initState();
-    initializeDateFormatting();
-    // DateTime.now().startOfDay
-    // DateTime.now().endOfDay
-    sportBookingService = BookingService(
-        // userEmail: ,
-        // userId: ,
-        serviceName: 'Sports Booking',
-        serviceDuration: 15,
-        bookingEnd: DateTime(now.year, now.month, now.day, 22, 00),
-        bookingStart: DateTime(now.year, now.month, now.day, 6, 30));
-  }
-
-  Stream<dynamic>? getBookingStreamMock(
-      {required DateTime end, required DateTime start}) {
-    return Stream.value([]);
-  }
-
-  Future<dynamic> uploadBookingMock(
-      {required BookingService newBooking}) async {
-    await Future.delayed(const Duration(seconds: 3));
-    converted.add(DateTimeRange(
-        start: newBooking.bookingStart, end: newBooking.bookingEnd));
-    if (kDebugMode) {
-      print('${newBooking.toJson()} has been uploaded');
-    }
-  }
-
-  List<DateTimeRange> converted = [];
-
-  List<DateTimeRange> convertStreamResultMock({required dynamic streamResult}) {
-    ///here you can parse the streamresult and convert to [List<DateTimeRange>]
-    ///take care this is only mock, so if you add today as disabledDays it will still be visible on the first load
-    ///disabledDays will properly work with real data
-    DateTime first = now;
-    // DateTime second = now.add(const Duration(minutes: 15));
-    // DateTime third = now.add(const Duration(minutes: 15));
-    // DateTime fourth = now.add(const Duration(minutes: 15));
-    // converted.add(
-    //     DateTimeRange(start: first, end: now.add(const Duration(minutes: 15))));
-    // converted.add(DateTimeRange(
-    //     start: second, end: second.add(const Duration(minutes: 15))));
-    // converted.add(DateTimeRange(
-    //     start: third, end: third.add(const Duration(minutes: 15))));
-    // converted.add(DateTimeRange(
-    //     start: fourth, end: fourth.add(const Duration(minutes: 15))));
-    return converted;
-  }
-
-  // List<DateTimeRange> generatePauseSlots() {
-  //   return [
-  //     DateTimeRange(
-  //         start: DateTime(now.year, now.month, now.day, 12, 0),
-  //         end: DateTime(now.year, now.month, now.day, 15, 0))
-  //   ];
-  // }
-
-  // String changeText(DateTime time) {
-  //   String timeChange = time.toString();
-  //   Text(timeChange) {
-  //     const TextStyle(fontSize: 15);
-  //   }
-
-  //   return timeChange;
-  // }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          backgroundColor: Colors.purple,
-          automaticallyImplyLeading: true,
-          title: const Text('Book a slot'),
-        ),
-        body: Padding(
-          padding: const EdgeInsets.fromLTRB(10.0, 10.0, 10.0, 85.0),
-          child: Center(
-            child: BookingCalendar(
-              bookingService: sportBookingService,
-              convertStreamResultToDateTimeRanges: convertStreamResultMock,
-              getBookingStream: getBookingStreamMock,
-              uploadBooking: uploadBookingMock,
-              hideBreakTime: true,
-              loadingWidget: const Text('Fetching data...'),
-              uploadingWidget:
-                  const CircularProgressIndicator(color: Colors.purple),
-              startingDayOfWeek: StartingDayOfWeek.monday,
-              locale: "en",
-              gridScrollPhysics:
-                  const ScrollPhysics(parent: AlwaysScrollableScrollPhysics()),
-              availableSlotText: 'Available',
-              selectedSlotText: 'Selected',
-              bookedSlotText: 'Booked',
-              // pauseSlotColor: Colors.grey,
-              bookingGridCrossAxisCount: 4,
-              bookingGridChildAspectRatio: 350 / 420,
-              // formatDateTime: changeText(),
-              // disabledDays: const [6, 7],
-            ),
-          ),
-        ));
   }
 }
