+++ lib/subscreens/climb.dart
@@ -1,0 +1,154 @@
+//Climbing Page
+import 'package:flutter/material.dart';
+
+class Climb extends StatefulWidget {
+  static const String route = '/Bookings/Climb';
+  const Climb({Key? key}) : super(key: key);
+
+  @override
+  State<Climb> createState() => ClimbState();
+}
+
+class ClimbState extends State<Climb> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(
+          title: const Text('Climbing'),
+          centerTitle: true,
+        ),
+        body: GridView(
+            padding: const EdgeInsets.all(20),
+            gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+              mainAxisSpacing: 15,
+              crossAxisCount: 2,
+              crossAxisSpacing: 15.0,
+            ),
+            children: [
+              // BOOKING CARD
+              InkWell(
+                onTap: () {},
+                child: Card(
+                  color: Colors.purple,
+                  shape: const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(15)),
+                  ),
+                  child: Column(
+                    children: [
+                      ClipRRect(
+                        borderRadius: const BorderRadius.only(
+                            topLeft: Radius.circular(15),
+                            topRight: Radius.circular(15)),
+                        child: Image.asset('lib/assets/images/climb_girl.png',
+                            height: 110,
+                            gaplessPlayback: true,
+                            fit: BoxFit.fill),
+                      ),
+                      const SizedBox(
+                        height: 10,
+                      ),
+                      const Text(
+                        'Climb booking',
+                        style: TextStyle(fontSize: 18, color: Colors.white),
+                      )
+                    ],
+                  ),
+                ),
+              ),
+
+              //SERVICES CARD
+              InkWell(
+                onTap: () {},
+                child: Card(
+                  color: Colors.purple,
+                  shape: const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(15)),
+                  ),
+                  child: Column(
+                    children: [
+                      ClipRRect(
+                        borderRadius: const BorderRadius.only(
+                            topLeft: Radius.circular(15),
+                            topRight: Radius.circular(15)),
+                        child: Image.asset('lib/assets/images/climb_3.png',
+                            height: 110,
+                            gaplessPlayback: true,
+                            fit: BoxFit.fill),
+                      ),
+                      const SizedBox(
+                        height: 10,
+                      ),
+                      const Text(
+                        'session help',
+                        style: TextStyle(fontSize: 18, color: Colors.white),
+                      )
+                    ],
+                  ),
+                ),
+              ),
+
+              //RULES CARD
+              InkWell(
+                onTap: () {},
+                child: Card(
+                  color: Colors.purple,
+                  shape: const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(15)),
+                  ),
+                  child: Column(
+                    children: [
+                      ClipRRect(
+                        borderRadius: const BorderRadius.only(
+                            topLeft: Radius.circular(15),
+                            topRight: Radius.circular(15)),
+                        child: Image.asset('lib/assets/images/climb_1.png',
+                            height: 110, width: 180, fit: BoxFit.fill),
+                      ),
+                      const SizedBox(
+                        height: 10,
+                      ),
+                      const Text(
+                        'Climbing \nEquipments',
+                        style: TextStyle(fontSize: 17, color: Colors.white),
+                      )
+                    ],
+                  ),
+                ),
+              ),
+
+              InkWell(
+                onTap: () {},
+                child: Card(
+                  color: Colors.purple,
+                  shape: const RoundedRectangleBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(15)),
+                  ),
+                  child: Column(
+                    children: [
+                      ClipRRect(
+                        borderRadius: const BorderRadius.only(
+                            topLeft: Radius.circular(15),
+                            topRight: Radius.circular(15)),
+                        child: Image.asset(
+                            'lib/assets/images/climb_2.png',
+                            height: 110,
+                            gaplessPlayback: true,
+                            fit: BoxFit.fill),
+                      ),
+                      const SizedBox(
+                        height: 10,
+                      ),
+                      const Text(
+                        'Under 16 \n bookings',
+                        style: TextStyle(
+                            fontSize: 17,
+                            color: Colors.white,
+                            fontWeight: FontWeight.normal),
+                      )
+                    ],
+                  ),
+                ),
+              ),
+            ]));
+  }
+}
